pyproject.toml
[project]
name = "MyTeam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "autogen-agentchat~=0.2",
    "datasets>=3.2.0",
    "litellm[proxy]>=1.55.2",
    "ollama>=0.4.4",
    "openai>=1.57.4",
    "pandas>=2.2.3",
    "tyro>=0.9.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
